refactor: Move release channel button to settings
