+++ src-vue/src/views/DeveloperView.vue
@@ -16,13 +16,6 @@
             Check NSProton Compatibility
         </el-button>
 
-        <h3>Testing:</h3>
-
-        <el-button type="primary" @click="toggleReleaseCandidate">
-            Toggle Release Candidate
-        </el-button>
-
-
         <h3>Mod install:</h3>
 
         <el-input v-model="mod_to_install_field_string" placeholder="Please input Thunderstore dependency string" clearable />
@@ -55,7 +48,6 @@
 import { defineComponent } from "vue";
 import { invoke } from "@tauri-apps/api";
 import { ElNotification } from "element-plus";
-import { ReleaseCanal } from "../utils/ReleaseCanal";
 import { GameInstall } from "../utils/GameInstall";
 import { Store } from 'tauri-plugin-store-api';
 const persistentStore = new Store('flight-core-settings.json');
@@ -99,28 +91,6 @@ export default defineComponent({
                     });
                     console.error(error);
                 });
-        },
-        async toggleReleaseCandidate() {
-            // Flip between RELEASE and RELEASE_CANDIDATE
-            this.$store.state.northstar_release_canal = this.$store.state.northstar_release_canal === ReleaseCanal.RELEASE
-                ? ReleaseCanal.RELEASE_CANDIDATE
-                : ReleaseCanal.RELEASE;
-
-            // Save change in persistent store
-            await persistentStore.set('northstar-release-canal', { value: this.$store.state.northstar_release_canal });
-
-            // Update current state so that update check etc can be performed
-            this.$store.commit("checkNorthstarUpdates");
-
-            console.log(this.$store.state)
-
-            // Display notification to highlight change
-            ElNotification({
-                title: `${this.$store.state.northstar_release_canal}`,
-                message: `Switched release channel to: "${this.$store.state.northstar_release_canal}"`,
-                type: 'success',
-                position: 'bottom-right'
-            });
         },
         async disableAllModsButCore() {
             let game_install = {
